integer safety: clamp gsize I/O counts; signed io_buf arithmetic
- Clamp the user's gsize read/write counts to INT_MAX before handing
  them to the int-based crypto I/O internals: a count > INT_MAX
  became a negative int (compiled-out assert), and crypto_read_low's
  loop could then write past the caller's buffer.  The plaintext path
  already clamped with SSIZE_MAX; partial I/O is valid GIOChannel
  behavior.
- Compute psl_io_buf sizes in signed arithmetic: on ILP32 (armv7) the
  guint/ssize_t mixes promoted to unsigned int, producing
  -Wsign-compare warnings and turning a broken invariant into a huge
  positive size instead of a detectably negative one.

+++ src/psl_io_buf.h
@@ -146,7 +146,10 @@ psl_io_buf_get_max_capacity(const PslIOBuf* const pIOBuf)
 PSL_CONFIG_INLINE_FUNC ssize_t
 psl_io_buf_get_data_size(const PslIOBuf* const pIOBuf)
 {
-    return (pIOBuf->garray_->len - pIOBuf->nextConsumeIndex_);
+    /// Compute in signed arithmetic so a broken invariant surfaces as a
+    /// detectably negative value instead of a huge unsigned one (the
+    /// operands would otherwise promote to unsigned int on ILP32)
+    return ((ssize_t)pIOBuf->garray_->len - pIOBuf->nextConsumeIndex_);
 }
 
 
@@ -320,8 +323,9 @@ psl_io_buf_extend_if_possible(PslIOBuf* const pIOBuf,
     ssize_t const maxRes = pIOBuf->maxSize_ - psl_io_buf_get_data_size(pIOBuf);
     ssize_t const actualRes = (reqCnt <= maxRes) ? reqCnt : maxRes;
 
-    if ((pIOBuf->garray_->len + actualRes) <= pIOBuf->maxSize_) {
-        (void)g_array_set_size(pIOBuf->garray_, pIOBuf->garray_->len + actualRes);
+    if (((ssize_t)pIOBuf->garray_->len + actualRes) <= pIOBuf->maxSize_) {
+        (void)g_array_set_size(pIOBuf->garray_,
+                               (guint)((ssize_t)pIOBuf->garray_->len + actualRes));
     }
     else {
         ssize_t oldDataSize;
